Split REST API integration files up
